requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
chromadb==0.4.24
sentence-transformers==2.5.1
semantic-router==0.0.14
numpy==1.26.4
pandas==2.2.1
openpyxl==3.1.2 